TAMAYA-326: Fix bug in ordinal comparison in ServiceContextManager
While adding some test coverage on the api, I found that
org.apache.tamaya.spi.ServiceContextManager has a bug as it searches for
the default service provider in loadDefaultServiceProvider.  Namely, the
"highestOrdinal" is not set after the first service provider is found,
so any following service provider with an ordinal higher than 0 will be
used, even if that ordinal is lower than the first one's.  This patch
fixes the logic, and adds tests to cover the case.